2011-01-24  Kai Tietz  <kai.tietz@onevision.com>
        * emit-rtl.c (reg_attrs_htab_hash): Replace long by intptr_t.

Additionally correct ChangeLog date entry for Koenig.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@169176 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/emit-rtl.c
@@ -326,7 +326,7 @@ reg_attrs_htab_hash (const void *x)
 {
   const reg_attrs *const p = (const reg_attrs *) x;
 
-  return ((p->offset * 1000) ^ (long) p->decl);
+  return ((p->offset * 1000) ^ (intptr_t) p->decl);
 }
 
 /* Returns nonzero if the value represented by X (which is really a
